refactor: onPressEntryDemoRoom move to service
