+++ src/screens/HomeScreen.tsx
@@ -2,10 +2,10 @@ import React, { useCallback, useEffect, useState } from 'react'
 import { StyleSheet, Dimensions, ScrollView, View } from 'react-native'
 import { NavigationStackProp, NavigationStackScreenProps } from 'react-navigation-stack'
 import { headerNavigationOptions } from '../navigators/options'
-import { useParties } from '../services/party'
+import { useParties, useEntryDemoRoom } from '../services/party'
 import { useModal } from '../services/modal'
 import { useStyles, MakeStyles } from '../services/design'
-import { useAuthState, useUIActions, useRoomActions } from '../store/hooks'
+import { useAuthState } from '../store/hooks'
 import { LoadingPage } from '../components/pages'
 import { Card, GenderModal, Modal } from '../components/organisms'
 import { setGender } from '../services/user'
@@ -17,10 +17,9 @@ type Props = OwnProps
 const HomeScreen = ({ navigation }: Props) => {
   const styles = useStyles(makeStyles)
   const { user, uid } = useAuthState()
-  const { entryDemoRoom } = useRoomActions()
-  const { openLoadingModal, closeLoadingModal } = useUIActions()
 
   const parties = useParties()
+  const { onPressEntryDemoRoom } = useEntryDemoRoom()
   const genderModalTools = useModal()
   const isAcceptedModalTools = useModal()
 
@@ -56,31 +55,30 @@ const HomeScreen = ({ navigation }: Props) => {
     }
   }, [genderModalTools, isSendGender, user])
 
-  // TODO: onPressEntryForDemoをservicesに退避させる。
-  const onPressEntryForDemo = useCallback(
-    async (party: Party) => {
-      if (!user || !user.gender) return
+  // const onPressEntryForDemo = useCallback(
+  //   (party: Party) => {
+  //     if (!user || !user.gender) return
 
-      if (!party.entryUIDs?.includes(user.uid)) {
-        openLoadingModal()
+  //     if (!party.entryUIDs?.includes(user.uid)) {
+  //       openLoadingModal()
 
-        const onSuccess = () => {
-          closeLoadingModal()
-          navigation.navigate('Chat', { roomID: party.id })
-        }
+  //       const onSuccess = () => {
+  //         closeLoadingModal()
+  //         navigation.navigate('Chat', { roomID: party.id })
+  //       }
 
-        const onFailure = () => {
-          closeLoadingModal()
-        }
+  //       const onFailure = () => {
+  //         closeLoadingModal()
+  //       }
 
-        entryDemoRoom({ roomID: party.id, onSuccess, onFailure })
-        return
-      }
+  //       entryDemoRoom({ roomID: party.id, onSuccess, onFailure })
+  //       return
+  //     }
 
-      navigation.navigate('Chat', { roomID: party.id })
-    },
-    [closeLoadingModal, entryDemoRoom, navigation, openLoadingModal, user]
-  )
+  //     navigation.navigate('Chat', { roomID: party.id })
+  //   },
+  //   [closeLoadingModal, entryDemoRoom, navigation, openLoadingModal, user]
+  // )
 
   const FetchPartiesThumbnail = useCallback(
     (parties: Party[]) => {
@@ -96,17 +94,17 @@ const HomeScreen = ({ navigation }: Props) => {
               date={party.date}
               width={width}
               onPressEntry={() => {
-                onPressEntryForDemo(party)
+                onPressEntryDemoRoom(party)
               }}
               // TODO: PartyEntryScreenで閉じれるものは閉じた方が良いので、PartyEntryScreenで定義して使用する。
-              onPressDetail={() => navigation.navigate('PartyDetail', { partyID, onPressEntry: onPressEntryForDemo })}
+              onPressDetail={() => navigation.navigate('PartyDetail', { partyID, onPressEntry: onPressEntryDemoRoom })}
             />
           </View>
         )
       })
       return thumbnailURLs
     },
-    [navigation, onPressEntryForDemo, styles.container]
+    [navigation, onPressEntryDemoRoom, styles.container]
   )
 
   if (!parties) {
